Review changes + Move pasted path stuff to clipboard_paste crate
